+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Unlockable Races of Tamriel (OpenMW)
 
+## 2.0.1
+
+- Slightly tweaked race list renderer's label padding
+
+## 2.0
+
+- Replaced settings renderer for displaying unlocked and locked races
+- Simplified mod logic
+- Changed default race unlocking conditions to more interesting ones
+
 ## 1.0.2
 
 - Removed leftover debug prints